+++ src/sld/bitmap.c
@@ -67,7 +67,7 @@ __sld_execute_bitmap(sld_entry *sld)
     gfx_dimensions screen;
     gfx_get_screen_dimensions(&screen);
 
-    uint16_t x, y;
+    int x, y;
     switch (sld->posx)
     {
     case SLD_ALIGN_CENTER:
@@ -77,12 +77,10 @@ __sld_execute_bitmap(sld_entry *sld)
         x = screen.width - bm.width;
         break;
     default:
-        x = (uint16_t)((int32_t)sld->posx * __sld_screen.width /
-                       SLD_VIEWBOX_WIDTH);
+        x = (int)((int32_t)sld->posx * __sld_screen.width / SLD_VIEWBOX_WIDTH);
     }
 
-    y = (uint16_t)((int32_t)sld->posy * __sld_screen.height /
-                   SLD_VIEWBOX_HEIGHT);
+    y = (int)((int32_t)sld->posy * __sld_screen.height / SLD_VIEWBOX_HEIGHT);
 
     if (!gfx_draw_bitmap(&bm, x, y))
     {
